+++ packages/video_player/video_player/test/video_player_test.dart
@@ -43,6 +43,9 @@ class FakeController extends ValueNotifier<VideoPlayerValue>
   Future<void> setLooping(bool looping) async {}
 
   @override
+  Future<void> setMuted(bool muted) async {}
+
+  @override
   Future<void> setAssetDataSource(
     String dataSource, {
     String package,
@@ -325,6 +328,17 @@ void main() {
       expect(controller.value.isLooping, isTrue);
     });
 
+    test('setMuted', () async {
+      final VideoPlayerController controller = VideoPlayerController.network(
+        'https://127.0.0.1',
+      );
+      await controller.initialize();
+      expect(controller.value.isMuted, isFalse);
+      await controller.setMuted(true);
+
+      expect(controller.value.isMuted, isTrue);
+    });
+
     test('pause', () async {
       final VideoPlayerController controller = VideoPlayerController();
       await controller.setNetworkDataSource('https://127.0.0.1');
@@ -383,6 +397,20 @@ void main() {
 
         await controller.setVolume(11);
         expect(controller.value.volume, 1.0);
+      });
+
+      test('disable mute', () async {
+        final VideoPlayerController controller = VideoPlayerController.network(
+          'https://127.0.0.1',
+        );
+        await controller.initialize();
+        await controller.setMuted(true);
+        expect(controller.value.isMuted, isTrue);
+
+        const double volume = 0.5;
+        await controller.setVolume(volume);
+
+        expect(controller.value.isMuted, isFalse);
       });
     });
 
@@ -481,6 +509,7 @@ void main() {
       expect(uninitialized.buffered, isEmpty);
       expect(uninitialized.isPlaying, isFalse);
       expect(uninitialized.isLooping, isFalse);
+      expect(uninitialized.isMuted, isFalse);
       expect(uninitialized.isBuffering, isFalse);
       expect(uninitialized.volume, 1.0);
       expect(uninitialized.errorDescription, isNull);
@@ -500,6 +529,7 @@ void main() {
       expect(error.buffered, isEmpty);
       expect(error.isPlaying, isFalse);
       expect(error.isLooping, isFalse);
+      expect(error.isMuted, isFalse);
       expect(error.isBuffering, isFalse);
       expect(error.volume, 1.0);
       expect(error.errorDescription, errorMessage);
@@ -520,6 +550,7 @@ void main() {
       const bool isPlaying = true;
       const bool isLooping = true;
       const bool isBuffering = true;
+      const bool isMuted = true;
       const double volume = 0.5;
 
       final VideoPlayerValue value = VideoPlayerValue(
@@ -530,10 +561,11 @@ void main() {
           isPlaying: isPlaying,
           isLooping: isLooping,
           isBuffering: isBuffering,
+          isMuted: isMuted,
           volume: volume);
 
       expect(value.toString(),
-          'VideoPlayerValue(duration: 0:00:05.000000, size: Size(400.0, 300.0), position: 0:00:01.000000, buffered: [DurationRange(start: 0:00:00.000000, end: 0:00:04.000000)], isPlaying: true, isLooping: true, isBuffering: truevolume: 0.5, errorDescription: null)');
+          'VideoPlayerValue(duration: 0:00:05.000000, size: Size(400.0, 300.0), position: 0:00:01.000000, buffered: [DurationRange(start: 0:00:00.000000, end: 0:00:04.000000)], isPlaying: true, isLooping: true, isBuffering: true, isMuted: true, volume: 0.5, errorDescription: null)');
     });
 
     test('copyWith()', () {
@@ -609,6 +641,7 @@ class FakeVideoPlayerPlatform {
       case 'pause':
       case 'play':
       case 'setLooping':
+      case 'setMuted':
       case 'setVolume':
         break;
       default: