+++ packages/video_player/video_player_web/lib/video_player_web.dart
@@ -96,6 +96,11 @@ class VideoPlayerPlugin extends VideoPlayerPlatform {
   }
 
   @override
+  Future<void> setMuted(int textureId, bool muted) async {
+    return _videoPlayers[textureId].setMuted(muted);
+  }
+
+  @override
   Future<void> seekTo(int textureId, Duration position) async {
     return _videoPlayers[textureId].seekTo(position);
   }
@@ -181,7 +186,12 @@ class _VideoPlayer {
   }
 
   void setVolume(double value) {
+    videoElement.muted = false;
     videoElement.volume = value;
+  }
+
+  void setMuted(bool muted) {
+    videoElement.muted = muted;
   }
 
   void seekTo(Duration position) {