+++ packages/video_player/video_player_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.0.2
+
+* Add `muted` attribute
+
 ## 2.0.1
 
 * Add caching functionality to videos from network sources.